+++ src/main/java/io/gittul/domain/post/PostRepository.java
@@ -1,0 +1,9 @@
+package io.gittul.domain.post;
+
+import io.gittul.domain.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
