+++ src/main/java/io/gittul/domain/post/entity/Post.java
@@ -1,17 +1,19 @@
 package io.gittul.domain.post.entity;
 
+import io.gittul.domain.bookmark.entity.Bookmark;
+import io.gittul.domain.comment.entity.Comment;
 import io.gittul.domain.github.entity.GitHubRepository;
-import io.gittul.domain.post.comment.entity.Comment;
-import io.gittul.domain.user.entity.Bookmark;
+import io.gittul.domain.like.enriry.UserLikePost;
+import io.gittul.domain.tag.entity.PostTag;
+import io.gittul.domain.tag.entity.Tag;
 import io.gittul.domain.user.entity.User;
 import io.gittul.global.jpa.EntityTimeStamp;
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import lombok.AccessLevel;
