+++ src/main/java/io/gittul/domain/tag/TagRepository.java
@@ -1,0 +1,12 @@
+package io.gittul.domain.tag;
+
+import io.gittul.domain.tag.entity.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface TagRepository extends JpaRepository<Tag, Long> {
+    List<Tag> findAllByTagName(String tagName);
+}
