+++ src/main/java/io/gittul/domain/user/entity/UserDetails.java
@@ -3,8 +3,7 @@ package io.gittul.domain.user.entity;
 import io.gittul.domain.post.comment.entity.Comment;
 import io.gittul.domain.post.comment.entity.UserLikeComment;
 import io.gittul.domain.post.entity.Post;
-import io.gittul.domain.post.entity.Tag;
-import io.gittul.domain.post.entity.UserLikePost;
+import io.gittul.domain.tag.entity.UserInterest;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
